12.04.2023

+++ src/main/java/com/anatoliiTelRan/rmtbanking/anatoliiCreditservice/entity/Agreement.java
@@ -31,7 +31,7 @@ public class Agreement {
 
 
     @Column(name = "status", nullable = false)
-    private Integer status;
+    private short status;
 
 
     @Column(name = "sum", nullable = false, precision = 2)
